+++ keyboards/microsoft-sculpt-mobile/Makefile
@@ -1,3 +1,6 @@
+SUBPROJECT_DEFAULT=teensy2pp
+
+
 ifndef MAKEFILE_INCLUDED
 	include ../../Makefile
 endif
