+++ keyboards/microsoft-sculpt-mobile/astar/Makefile
@@ -1,0 +1,3 @@
+ifndef MAKEFILE_INCLUDED
+	include ../../../Makefile
+endif
