+++ keyboards/microsoft-sculpt-mobile/astar/readme.md
@@ -1,0 +1,29 @@
+The Astar mini has all pins exposed , so you can do 18x8
+If you want a speaker, LEDs &etc, you'll need to free up a pin. 
+
+Here's the preferred shrinking to 17 Cols + 8 Rows for 32u4 boards. 
+Connect both columns R and L  to the same pin. 
+- Note that Rshift+Tab and FN+Rshift, FN+x and FN+2 are on same column and will ghost. 
+Other candidates are Col K or Col N, but I want to be able to chord RAlt and R-Ctl with FN. 
+    			
+    			
+        A       B       C       D       E       F       G       H  --->
+0       b       n       m       ,       .       /                  --->
+1       g       h       "                                          --->          
+2       7       8       9       0       Del             PgUp       --->          
+3       p       [       ]       \                                  --->          
+4       y       u       i       o                                  --->               
+5       ~       -       +=      j       k       l       ;       5  --->               
+6       a       s       d       q       w       e,      Up      left-->     
+7       F7      F8      F9      F10     F11     F12             f  --->                  
+    			
+------->    I       J       K       L(+R)   M      N       O       P       Q    
+--->0               Caps           FN                                             
+--->1                      Vol+    mute    Rctl            vol-    pgdn    LCTL            
+--->2                              Rshift                          LShift                  
+--->3                      Ralt                            LAlt                    
+--->4                                              LGUI                                            
+--->5       6       bakspc  1       2       3       4       F4      F5      F6                      
+--->6       Down    right           Tab            spc      F1      F2      F3           
+--->7       r       t       z       x       c               v       enter   Esc                     
+
