+++ keyboards/microsoft-sculpt-mobile/teensy2pp/Makefile
@@ -1,0 +1,3 @@
+ifndef MAKEFILE_INCLUDED
+	include ../../../Makefile
+endif
