+++ keyboards/microsoft-sculpt-mobile/teensy2pp/readme.md
@@ -1,0 +1,25 @@
+
+The full original keymap for the sculpt is 
+        A       B       C       D       E       F       G       H   --->  
+0       b       n       m       ,       .       /                
+1       g       h       "                                          
+2       7       8       9       0       Del             PgUp     
+3       p       [       ]       \                                
+4       y       u       i       o                                      
+5       ~       -       +=      j       k       l       ;       5
+6       a       s       d       q       w       e,      Up      left    
+7       F7      F8      F9      F10     F11     F12             f
+    			
+----->  I       J       K       L       M       N       O       P       Q       R       
+0               Caps                                                            FN      
+1                       Vol+    mute    Rctl            vol-    pgdn    LCTL            
+2                               Rshift                          LShift                  
+3                       Ralt                            LAlt                    
+4                                               LGUI                                            
+5       6       bakspc  1       2       3       4       F4      F5      F6                      
+6       Down    right                           spc     F1      F2      F3      tab             
+7       r       t       z       x       c               v       enter   Esc                     
+
+This works with 18 cols + 8  rows on a Teensy++, or Arm based Teensy. 
+
+
